+++ examples/Unit_KMeter_M5Device/simple/simple.ino
@@ -22,7 +22,7 @@ M5_KMeter sensor;
 void setup(void) {
     display.begin();
 
-    Wire.begin(SDA, SCL, 400000L);
+    Wire.begin((int)SDA, (int)SCL, 400000L);
     // Wire.begin(21, 22, 400000L);
 
     sensor.begin();
